backend/requirements.txt
# FastAPI核心依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0



# 任务队列
celery>=5.3.0
redis>=5.0.0

# 工具
python-dotenv>=1.0.0
pytz>=2023.3

# PDF生成
reportlab>=4.0.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

pandas>=2.0.3
numpy>=1.24.3
plotly>=5.15.0
yfinance>=0.2.18
akshare>=1.11.0
tushare>=1.3.0
openai>=1.12.0
python-dotenv>=1.0.0
pytz
ta>=0.10.2
reportlab>=4.0.0
peewee>=3.17.0
schedule>=1.2.0 
pywencai>=0.7.0
